relocate robot related states from teleoperate and add some styling
