+++ lerobot-web-fe/src/components/Teleoperate.module.css
@@ -1,0 +1,85 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+  padding: 1rem 0rem;
+  font-family: 'Segoe UI', 'Helvetica Neue', sans-serif;
+}
+
+.statusBox {
+  background-color: #f9f9fb;
+  padding: 1rem 1.5rem;
+  border-radius: 8px;
+}
+
+.statusTitle {
+  font-size: 1.25rem;
+  font-weight: 600;
+  margin: 0 0 0.5rem;
+  color: #333;
+}
+
+.statusText {
+  font-size: 1rem;
+  color: #555;
+}
+
+.errorText {
+  color: #d32f2f;
+  font-size: 0.95rem;
+  margin-top: 0.5rem;
+}
+
+.controlButton {
+  display: inline-flex;
+  align-items: center;
+  width: fit-content;
+  gap: 0.5rem;
+  border: none;
+  font-size: 1rem;
+  font-weight: 600;
+  padding: 0.75rem 1.25rem;
+  margin-left: 10px;
+  border-radius: 6px;
+  cursor: pointer;
+  transition: background 0.3s ease;
+}
+
+.start {
+  background-color: #3c91ff;
+  color: white;
+}
+
+.start:hover {
+  background-color: #1f78e0;
+}
+
+.stop {
+  background-color: #ff4f4f;
+  color: white;
+}
+
+.stop:hover {
+  background-color: #d93939;
+}
+
+.icon {
+  width: 1rem;
+  height: 1rem;
+}
+
+.loader {
+  border: 2px solid transparent;
+  border-top: 2px solid white;
+  border-right: 2px solid white;
+  border-radius: 50%;
+  width: 0.9rem;
+  height: 0.9rem;
+  animation: spin 0.8s linear infinite;
+}
+
+@keyframes spin {
+  to {
+    transform: rotate(360deg);
+  }
+}
